MOR-5: adapt RatingEntry value to Double, create RatingEntryRepository

+++ src/main/java/en/ratings/own/my/model/rating/RatingEntry.java
@@ -29,9 +29,9 @@ public class RatingEntry {
     private String name;
 
     @NonNull
-    private Long value;
+    private Double value;
 
-    public RatingEntry(Long ratingId, String name, Long value) {
+    public RatingEntry(Long ratingId, String name, Double value) {
         setRatingId(ratingId);
         setName(name);
         setValue(value);
